Set up environments as a separate requireable gem within the project
