+++ Gemfile
@@ -1,7 +1,7 @@
 source "https://rubygems.org"
 
-# Specify your gem's dependencies in retriable.gemspec
-gemspec
+gemspec name: 'retriable'
+gemspec name: 'retriable_environments'
 
 group :test do
   # gem "ruby_gntp"
