+++ lib/retriable/exponential_backoff.rb
@@ -25,6 +25,7 @@ module Retriable
     end
 
     private
+
     def randomize(interval)
       return interval if rand_factor == 0
       delta = rand_factor * interval * 1.0
