+++ lib/retriable/version.rb
@@ -1,3 +1,3 @@
 module Retriable
-  VERSION = "3.1.0"
+  VERSION = '3.1.0'
 end
