+++ lib/retriable_environments/version.rb
@@ -1,0 +1,3 @@
+module RetriableEnvironments
+  VERSION = '0.0.1'
+end
